move start logic into entry builder
